proyecto/requirements.txt
django
djangorestframework
django-filter
pillow